Crate ic_crypto_tls

Offers TLS functionality with self-managed keys.

As opposed to the CryptoComponent that manages the TLS private keys itself, this module allows the caller to manage the private keys. This can e.g. be used by parties that need to connect to Internet Computer nodes via TLS.

Security Warning

Since the private keys are self-managed, it is the responsibility of the caller to keep the private key material secure!

Structs

TlsEncodingError
A TLS struct couldn't be encoded (as PEM or DER).

TlsPemParsingError
A PEM string could not be parsed.

TlsPrivateKey
TLS private key

TlsPublicKeyCert
An X.509 certificate

Enums

TlsClientHandshakeError
The TLS client handshake failed.

Functions

generate_tls_keys
Generates and returns a private key and a self-signed X509 public key certificate.

perform_tls_client_handshake
Transforms a TCP stream into a TLS stream by performing a TLS client handshake.