Function ic_crypto_tls::perform_tls_client_handshake

pub async fn perform_tls_client_handshake(
    tcp_stream: TcpStream, 
    client_cert: TlsPublicKeyCert, 
    client_private_key: TlsPrivateKey, 
    trusted_server_cert: TlsPublicKeyCert
) -> Result<TlsStream, TlsClientHandshakeError>

Transforms a TCP stream into a TLS stream by performing a TLS client handshake.

This client authenticates its connection using the client_cert and the corresponding client_private_key. The client only connects if the server presents trusted_server_cert as certificate during the handshake.

For the handshake, the client uses the following configuration:

Minimum protocol version: TLS 1.3

Supported signature algorithms: ed25519

Allowed cipher suites: TLS_AES_128_GCM_SHA256, TLS_AES_256_GCM_SHA384

Server authentication: mandatory, with ed25519 certificate

The given tcp_stream is consumed. If an error is returned, the TCP connection is therefore dropped.

Arguments

tcp_stream is the TCP connection over which to run the TLS handshake

client_cert is the caller's X.509 certificate

client_private_key is the private key corresponding to client_cert

trusted_server_cert is the X.509 certificate of the intended peer

Returns

Ok(TlsStream) if the handshake was successful, TlsClientHandshakeError otherwise

Errors

TlsClientHandshakeError::CreateConnectorError if the key or certs fail to parse or there is an error creating the configuration

TlsClientHandshakeError::HandshakeError if the handshake fails